% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_finder.R
\name{cv_groups}
\alias{cv_groups}
\title{Defines environment groups for k-fold cross validation}
\usage{
cv_groups(E, folds = 6)
}
\arguments{
\item{E}{A character vector of environments included in the multi-environment trial data}

\item{folds}{The number of fold in the k-fold cross validation scheme}
}
\value{
A data frame with 2 columns. The first column is the environment term. The 2nd column is the cross validation group that each environment hsa been randomly allocated to.
}
\description{
Generates the cross validation groups for each environment to go into. These groups are then used for the environmental covariate selection procedure.
}
