% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropFixedTerm}
\alias{dropFixedTerm}
\title{Determine if corresponding spline term is present for each linear term}
\usage{
dropFixedTerm(.fm, wald_df, .ecs_in_model, .M)
}
\arguments{
\item{.fm}{An \code{asreml} model object}

\item{wald_df}{A data frame containing the information pertaining to the wald test (see \code{asreml::wald.asreml})}

\item{.ecs_in_model}{Need to check (quosure?)}

\item{.M}{A character indicating which term in the model is the management practice component}

\item{randomTerms}{Check if this is still needed in the model}
}
\value{
An updated Wald table with an additional column of type logical indicating whether each fixed effect term in the model can be dropped.
}
\description{
A function to determine if the corresponding spline term for a fixed effects EC term is currently in the model.
If the spline term is still in the model, then the linear term (fit as a fixed effect) cannot be dropped from the model, and hence will not be tested in \code{ec_fixed_model}.
}
