% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ev_cv.R
\name{ec_all}
\alias{ec_all}
\title{Include all important ECs into the model}
\usage{
ec_all(
  fm,
  ECs,
  G,
  E,
  M,
  env_cv_df = NULL,
  ncores = 2,
  kn = 6,
  trial = NULL,
  ecs_in_bline_model = rlang::maybe_missing()
)
}
\arguments{
\item{fm}{The baseline \code{asreml} model object that environmental covariates will be added to.
Note that the data frame used in the baseline model will be the dataframe used to identify each of the corresponding terms in the model.}

\item{ECs}{An expression with the environmental covariate to be included in the model.}

\item{G}{The genotype term in the model as an expression}

\item{E}{The environment term in the model as an expression}

\item{M}{The management practice term in the model as an expression}

\item{env_cv_df}{A data frame identifying which cross-validation group each environment belongs to.
If not provided, the environments will be randomly allocated to cross validation by calling \code{cv_groups} internally}

\item{ncores}{The number of computer cores used during the cross validation scheme. Note that this should not be more than the number of cross validation groups in the model.
Also note that by setting this greater than 2 (i.e. the default) will require additional \code{ASReml-R} licenses.}

\item{kn}{The number of knot points for the spline component of the model for the environment covariate being tested.
Note that this is ignored of the environmental covariate is a factor}

\item{trial}{The trial term  in the model as an expression. \code{.trial} is only required when each environment is not uniquely define by a trial.
An example is when each TOS within a trial constitutes an environment for the purposes of identifying important environmental covariates}

\item{ecs_in_bline_model}{An optional list of quosures where each element of the list pertains to an environmental covariate that is in the initial baseline model.}
}
\value{
A list with 2 components:
\itemize{
\item \code{fm}: An updated model with the all of the important environmental covariate included in the model along with only the significant
fixed and random effects pertaining to each environmental covariate identified.
\item \code{rmse}: The value of the RMSE for the final environmental covariate model
}
}
\description{
This function combines all of the component functions required to complete the environment covariate selection algorithm for a given baseline model.
The final output is an updated model containing each of the important ECs identified by the algorithm, along with the corresponding
environmental covariate terms that are statistically significant.
}
