% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_finder.R
\name{ec_finder}
\alias{ec_finder}
\title{Include multiple ECs into the model one at a time}
\usage{
ec_finder(
  fm,
  ECs,
  G,
  E,
  M,
  env_cv_df = NULL,
  cores = 2,
  kn = 6,
  trial = NULL,
  ecs_in_bline_model = NULL
)
}
\arguments{
\item{fm}{The baseline \code{asreml} model object that environmental covariates will be added to.
Note that the data frame used in the baseline model will be the dataframe used to identify each of the corresponding terms in the model.}

\item{ECs}{An expression with the environmental covariate to be included in the model.}

\item{G}{The genotype term in the model as an expression}

\item{E}{The environment term in the model as an expression}

\item{M}{The management practice term in the model as an expression}

\item{env_cv_df}{A data frame identifying which cross-validation group each environment belongs to.
If not provided, the environments will be randomly allocated to cross validation by calling \code{cv_groups} internally}

\item{cores}{The number of computer cores used during the cross validation scheme. Note that this should not be more than the number of cross validation groups in the model.
Also note that by setting this greater than 2 (i.e. the default) will require additional \code{ASReml-R} licenses.}

\item{kn}{The number of knot points for the spline component of the model for the environment covariate being tested.
Note that this is ignored of the environmental covariate is a factor}

\item{trial}{The trial term  in the model as an expression. \code{.trial} is only required when each environment is not uniquely define by a trial.
An example is when each TOS within a trial constitutes an environment for the purposes of identifying important environmental covariates}

\item{ecs_in_bline_model}{An optional list of quosures where each element of the list pertains to an environmental covariate that is in the initial baseline model.}
}
\value{
A list with components:
\itemize{
\item \code{ec_selected}: The environmental covariate seected as minimising the RMSE (a scalar of type character).
compared with the predictions obtained with the environment covariate included in an untested environment.
The predictions are obtained internally using \code{asreml::predict.asreml}.
\item \code{summary_ecs}: A data frame with the RMSE and Pearson correlation for each environmental covariate considered during the forward selection procedure.
}
}
\description{
This function implements the forward selection procedure for each environmental covariates (EC).
Each EC is incorporate into the model individually and then k-fold cross validation is performed.
Predcitions are then obtained in an untested environment, which are then compared back to the baseline model to obtain an RMSE for each environment.
The EC which minimises the RMSE is selected from the forward selection procedure.
If management practice \code{.M} and the environment covariates \code{.ec} are factors, then the predictions will be for each unique combination of \code{GxExM}.
If \code{.M} is continuous, then predictions will be very for every unique \code{.M} value observed in the multi-environment trial data.
}
