% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_ec_model.R
\name{ec_fixed_model}
\alias{ec_fixed_model}
\title{Identify the most parsimonious EC fixed effects model}
\usage{
ec_fixed_model(.fm, .ecs_in_model, .G, .M)
}
\arguments{
\item{.fm}{The current \code{asreml} model object that the environmental covariate terms will be tested from.}

\item{.ecs_in_model}{A list of expressions such that each expression is an environmental covariate that is present in the current model.}

\item{.G}{The genotype term in the model as a character}

\item{.M}{The management practice term in the model a character}
}
\value{
An updated asreml model such that the non-significant environmental covariate fixed effect terms have been removed from the model.
}
\description{
Performs the backwards selection procedure for the fixed component of the model to obtain a parsimonious fixed effects model,
removing all non-significant environmental covariate terms.
Testing for the significance of fixed effect terms is performed using Wald tests with an approximate F-statistic (Kenward & Roger 1997).
}
