% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ec_full_model_constructor}
\alias{ec_full_model_constructor}
\title{Creates an updated model which is the same as the current model but adds one new EC into the model}
\usage{
ec_full_model_constructor(.fm, .ec, .G, .M, .kn = 6)
}
\arguments{
\item{.fm}{An \code{asreml} model object without the environmental covariate in the model.}

\item{.ec}{An expression with the environmental covariate to be included in the model.}

\item{.G}{A character indicating which term in the model is the genotype component.}

\item{.M}{A character indicating which term in the model is the management practice component.}

\item{.kn}{The number of knot points to be included for the spline terms related to the environmental covariate \code{.ec} being added to the model.}
}
\value{
An asreml model object with all of the fixed and random terms related to the environmental covariate \code{.ec} being added to the model.
}
\description{
Runs an updated asreml model with the additional fixed and random effects term for the environmental covariate added to the model.
}
