% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ev_cv.R
\name{ec_iteration}
\alias{ec_iteration}
\title{Select the best EC and the significant terms}
\usage{
ec_iteration(
  fm,
  ECs,
  G,
  E,
  M,
  env_cv_df = NULL,
  ncores = 2,
  kn = 6,
  trial = NULL,
  ecs_in_bline_model = rlang::maybe_missing(),
  denDF = "numeric"
)
}
\arguments{
\item{fm}{The baseline \code{asreml} model object that environmental covariates will be added to.
Note that the data frame used in the baseline model will be the dataframe used to identify each of the corresponding terms in the model.}

\item{ECs}{An expression with the environmental covariate to be included in the model.}

\item{G}{The genotype term in the model as an expression}

\item{E}{The environment term in the model as an expression}

\item{M}{The management practice term in the model as an expression}

\item{env_cv_df}{A data frame identifying which cross-validation group each environment belongs to.
If not provided, the environments will be randomly allocated to cross validation by calling \code{cv_groups} internally}

\item{ncores}{The number of computer cores used during the cross validation scheme. Note that this should not be more than the number of cross validation groups in the model.
Also note that by setting this greater than 2 (i.e. the default) will require additional \code{ASReml-R} licenses.}

\item{kn}{The number of knot points for the spline component of the model for the environment covariate being tested.
Note that this is ignored of the environmental covariate is a factor}

\item{trial}{The trial term  in the model as an expression. \code{.trial} is only required when each environment is not uniquely define by a trial.
An example is when each TOS within a trial constitutes an environment for the purposes of identifying important environmental covariates}

\item{ecs_in_bline_model}{An optional list of quosures where each element of the list pertains to an environmental covariate that is in the initial baseline model.}

\item{denDF}{A character outlining the method used by \code{asreml} to calculate the denominator degrees of freedom.
Options include \code{none}, \code{algebraic} or \code{numeric}.}
}
\value{
A list with 2 components:
\itemize{
\item \code{fm}: An updated model with the most important environmental covariate included in the model along with only the significant
fixed and random effects pertaining to the environmental covariate identified.
\item \code{rmse}: The value of the RMSE for the updated model.
}
}
\description{
This function bring together the entire subset selection process to (i) identify the most important environmental covariate
via the forward selection procedure and then drop the non-significant fixed and random terms from the model to achieve a parsimonious model.
}
\examples{
library(asreml)
library(foreach)
data(SorghumYield)
data(SorghumCvGroup)
# Run baseline model
baseline_asr <- asreml::asreml( Yld ~ Genotype + density + Genotype:density,
  random =~ at(Trial):Rep  + at(Trial):MainPlot +
   at(Trial,c('Breeza 1', 'Breeza 2', 'Emerald', 'Moree')):SubPlot +
   at(Trial,'Breeza 1'):Column +
   Trial + Env +
   spl(density, k=6) + spl(density, k=6):Genotype +
   str(~Trial:Genotype + Trial:Genotype:density,
       ~corh(2):id(48)) +
   str(~Env:Genotype + Env:Genotype:density,
       ~corh(2):id(136)),
  residual=~ dsum(~units|ResidualTOS),
  data = SorghumYield,
  na.action=na.method(x='include'),
  maxit=30, workspace="1Gb")
  ec1_model_asr <- ec_iteration(fm=baseline_asr, ECs = rlang::quos(PrePAW:PostPAW, PreCumRad),
      G= rlang::expr(Genotype), E = rlang::expr(Env),
      M= rlang::expr(density), trial = rlang::expr(Trial),
      env_cv_df=SorghumCvGroup)
ec1_model_asr$selected_ec
ec1_model_asr$fm$call
ec1_model_asr$rmse
}
