% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expr_to_terms}
\alias{expr_to_terms}
\title{Convert an expression into an asreml call term}
\usage{
expr_to_terms(expr)
}
\arguments{
\item{expr}{An expression}
}
\value{
A character vector
}
\description{
A function to converts an expression into the appropriate term to go into an asreml call
}
