% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{margin}
\alias{margin}
\title{Determine the margin (i.e. order of hierarchy)}
\usage{
margin(terms, ec, G, M)
}
\arguments{
\item{terms}{A character vector of terms in the asreml model}

\item{ec}{A character indicating which term in the model is the environmental covariate being tested}

\item{G}{A character indicating which term in the model is the genotype component}

\item{M}{A character indicating which term in the model is the management practice component}
}
\value{
A vector of type double indicating the hierarchical structure of the random terms in the model for a particular environmental covariate
}
\description{
This function calculates the equivalent of the margin terms (used in the wald tests for fixed effects) for the random terms in the model,
enabling ec_random_model to correctly determine which terms to test and when to finishing testing for significant interaction terms.
}
