% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parallel_predict_list}
\alias{parallel_predict_list}
\title{Creates the list of information required for asreml.predict with continuous variables}
\usage{
parallel_predict_list(
  .df,
  subset_df,
  .ec = NULL,
  .G,
  .E,
  .M,
  baseline_ec_cols = NULL
)
}
\arguments{
\item{.df}{The data frame used in the core asreml model}

\item{subset_df}{The data frame used in the subbsetted version during the cross validation procedure}

\item{.ec}{An expression with the environmental covariate to be included in the model.}

\item{.G}{An expression indicating which term in the model is the genotype component.}

\item{.M}{An expression indicating which term in the model is the management practice component.}

\item{baseline_ec_cols}{A vector of integers indicating which columns in both \code{.df} and \code{subset_df} contain
environmental covariates that are in the current model}
}
\value{
A list of information required for asreml.predict with continuous variables (TBC!)
}
\description{
Creates the list of information required for asreml.predict with continuous variables
}
