% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_ec_model.R
\name{simplify_ec_model}
\alias{simplify_ec_model}
\title{Identify the most parsimonious EC model}
\usage{
simplify_ec_model(.fm, .ecs_in_model, .G, .M)
}
\arguments{
\item{.fm}{The current \code{asreml} model object that the environmental covariate terms will be tested from.}

\item{.ecs_in_model}{A list of expressions such that each expression is an environmental covariate that is present in the current model.}

\item{.G}{The genotype term in the model as a character}

\item{.M}{The management practice term in the model a character}
}
\value{
An updated asreml model such that all non-significant environmental covariate terms have been removed from the model.
}
\description{
Performs the backwards selection procedure for all environmental covariates in the curent model.
This is achieve by cyclcing between dropping fixed and random environmental covariate terms until all non-significant environmental
covariate terms have been dropped from the current model.
}
\examples{
library(asreml)
data(SorghumYield)
data(SorghumCvGroup)
# Run baseline model
baseline_asr <- asreml( Yld ~ Genotype + density + Genotype:density,
  random =~ at(Trial):Rep  + at(Trial):MainPlot +
   at(Trial,c('Breeza 1', 'Breeza 2', 'Emerald', 'Moree')):SubPlot +
   at(Trial,'Breeza 1'):Column +
   Trial + Env +
   spl(density, k=6) + spl(density, k=6):Genotype +
   str(~Trial:Genotype + Trial:Genotype:density,
       ~corh(2):id(48)) +
   str(~Env:Genotype + Env:Genotype:density,
       ~corh(2):id(136)),
  residual=~ dsum(~units|ResidualTOS),
  data = SorghumYield,
  na.action=na.method(x='include'),
  maxit=30, workspace="1Gb")
postPAW_full_asr <-  ec_full_model_constructor(.fm=baseline_asr, .ec=rlang::expr(PostPAW), .G=rlang::expr(Genotype), .M=rlang::expr(density))
simplify_asr <- simplify_ec_model(.fm=postPAW_full_asr, .ecs_in_model=rlang::quos(PostPAW), .G="Genotype", .M="density")
simplify_asr$call
}
