% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subtract_terms}
\alias{subtract_terms}
\title{Subtract one expression from a character vector}
\usage{
subtract_terms(main_expr, removed_char_vec, response = FALSE)
}
\arguments{
\item{main_expr}{An expression}

\item{removed_char_vec}{A character vector? (Need to double check)}

\item{response}{A logical determining whether the response variable is included in main_expr. Default is FALSE.}
}
\value{
An expression to be included in an asreml call.
}
\description{
A function that enables the subtraction of one character from a character vector. Used to ensure that the correct asreml model is called.
}
