% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_fixed_asr}
\alias{update_fixed_asr}
\title{Run an asreml model with updated fixed effects component}
\usage{
update_fixed_asr(.fm, term = rlang::maybe_missing(), ssType = "conditional")
}
\arguments{
\item{.fm}{An \code{asreml} model object}

\item{term}{A character scalar indicating which term in the model is currently being tested}

\item{ssType}{The type of sums of squares used for testing. Default is "\code{conditional}" which is analogous to Type 3 sums of squares.
Use "\code{incremental}" for Type 1 sums of squares. See  \code{asreml::wald.asreml} for more information.}
}
\value{
An asreml model object with a new fixed effets model.
}
\description{
This function is used to runs an updated version of asreml with a modified fixed effects model within \code{ec_fixed_model} with higher order fixed terms dropped in the model.
This is necessary for regression modelling to ensure that the Wald tests are correct for lower order terms in the model.
}
