% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ev_cv.R
\name{rmse_calc}
\alias{rmse_calc}
\title{Calculate the RMSE for a given model}
\usage{
rmse_calc(
  .fm,
  .G,
  .E,
  .M,
  .trial = NULL,
  .env_cv_df = NULL,
  .cores = 2,
  .kn = 6,
  .ecs_in_bline_model = rlang::quos(NULL)
)
}
\arguments{
\item{.fm}{The baseline \code{asreml} model object that environmental covariates will be added to.
Note that the data frame used in the baseline model will be the dataframe used to identify each of the corresponding terms in the model.}

\item{.G}{The genotype term in the model as an expression}

\item{.E}{The environment term in the model as an expression}

\item{.M}{The management practice term in the model as an expression}

\item{.trial}{The trial term  in the model as an expression. \code{.trial} is only required when each environment is not uniquely define by a trial.
An example is when each TOS within a trial constitutes an environment for the purposes of identifying important environmental covariates}

\item{.env_cv_df}{A data frame identifying which cross-validation group each environment belongs to.
If not provided, the environments will be randomly allocated to cross validation by calling \code{cv_groups} internally}

\item{.cores}{The number of computer cores used during the cross validation scheme. Note that this should not be more than the number of cross validation groups in the model.
Also note that by setting this greater than 2 (i.e. the default) will require additional \code{ASReml-R} licenses.}

\item{.kn}{The number of knot points for the spline component of the model for the environment covariate being tested.
Note that this is ignored of the environmental covariate is a factor}

\item{.ecs_in_bline_model}{An optional list of quosures where each element of the list pertains to an environmental covariate that is in the initial baseline model.}
}
\value{
A list with components:
\itemize{
\item \code{Cor}: The Pearson correlation of the predicted values from the baseline model
compared with the predictions obtained with the environment covariate included in an untested environment.
The predictions are obtained internally using \code{asreml::predict.asreml}.
\item \code{Rmse}: the full log-likelihood for each model
\item \code{Resids_cv}: The squared differences between the baseline model and the cross-validation predictions for each \code{GxExM} combination
}
}
\description{
This function calculates the RMSE for the current model by implementing the k-fold cross validation and then calculating the
predictions obtained in an untested environment.
}
\examples{
library(asreml)
data(SorghumYield)
data(SorghumCvGroup)
# Run baseline model
baseline_asr <- asreml( Yld ~ Genotype + density + Genotype:density,
  random =~ at(Trial):Rep  + at(Trial):MainPlot +
   at(Trial,c('Breeza 1', 'Breeza 2', 'Emerald', 'Moree')):SubPlot +
   at(Trial,'Breeza 1'):Column +
   Trial + Env +
   spl(density, k=6) + spl(density, k=6):Genotype +
   str(~Trial:Genotype + Trial:Genotype:density,
       ~corh(2):id(48)) +
   str(~Env:Genotype + Env:Genotype:density,
       ~corh(2):id(136)),
  residual=~ dsum(~units|ResidualTOS),
  data = SorghumYield,
  na.action=na.method(x='include'),
  maxit=30, workspace="1Gb")
# Calculate the (average) root mean square error (rmse) of prediction
baseline_rmse <- rmse_calc(.fm=baseline_asr, .G=rlang::expr(Genotype), .E=rlang::expr(Env), .M=rlang::expr(density),
                  .trial=rlang::expr(Trial), .env_cv_df=SorghumCvGroup)
baseline_rmse$Rmse
}
