% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_ec_model.R
\name{simplify_ec_model}
\alias{simplify_ec_model}
\title{Identify the most parsimonious EC model}
\usage{
simplify_ec_model(.fm, .ecs_in_model, .G, .M)
}
\arguments{
\item{.fm}{The current \code{asreml} model object that the environmental covariate terms will be tested from.}

\item{.ecs_in_model}{A list of expressions such that each expression is an environmental covariate that is present in the current model.}

\item{.G}{The genotype term in the model as a character}

\item{.M}{The management practice term in the model a character}
}
\value{
An updated asreml model such that all non-significant environmental covariate terms have been removed from the model.
}
\description{
Performs the backwards selection procedure for all environmental covariates in the curent model.
This is achieve by cyclcing between dropping fixed and random environmental covariate terms until all non-significant environmental
covariate terms have been dropped from the current model.
}
